Renamed article_controls ArticleControls
