+++ packages/frontend-web/src/app/components/index.ts
@@ -45,5 +45,5 @@ export * from './TagLabelRow';
 export * from './Toast';
 export * from './Toggle';
 export * from './ToolTip';
-export * from './article_controls';
+export * from './ArticleControls';
 export * from './HeaderBar';
